Functional geko.c
