+++ include/geko.h
@@ -1,6 +1,9 @@
 //
 //
 //
+/****************************************/
+// Libraries to include
+/****************************************/
 
 #include <stdio.h>   // for printf()
 #include <unistd.h>  // for getopt()
@@ -14,39 +17,14 @@
 #include "ini.h"     // for INI parsing
 #include "ini.c"     // for INI parsing
 
-// CONSTANTS
+/****************************************/
+// Constants
+/****************************************/
+
 #define FILESUFFIX ".dat"  // suffix for the output data file
 #define VER "0.1.0"
 #define GREET "geko v" VER " - geKo electrophysiology Kommander"
 #define DAQmxErrChk(functionCall) if( DAQmxFailed(error=(functionCall)) ) goto Error; else
-
-/*********************************************/
-// DAQmx Configuration Options
-/*********************************************/
-
-float64 minVal = -10.0; // The minimum value, in units, that you expect to generate.
-float64 maxVal = 10.0; // The maximum value, in units, that you expect to generate.
-
-float64 timeout = -1; // The amount of time, in seconds, to wait for the function to read the sample(s).
-
-// TODO: NOT RESPONDING
-bool32 dataLayout = DAQmx_Val_GroupByScanNumber; // Specifies how the samples are arranged, either interleaved or noninterleaved. Options: DAQmx_Val_GroupByChannel, DAQmx_Val_GroupByScanNumber
-
-// GLOBAL VARIABLES (initialized to zero)
-unsigned int reps = 1;   // number of repetitions of the stimulation protocol
-unsigned int isi = 0;   // inter-stimulus interval in seconds
-char *suffix = "";    // suffix for the output data file
-
-char* configFile = "../geko.ini";  // name of the configuration file
-char *stimFile = "default.stim";    // name of the stimulation protocol file
-char * FileName = NULL;     // name of the output data file to be created
-char *user_prompt = NULL;   // command used to run this program
-int ai_factor = 1;
-double ao_factor = 1.0;
-
-//float64 data[];
-
-// INI parsing
 typedef struct
 {
     int srate;
@@ -61,25 +39,58 @@ typedef struct
     double ao_factor_vc;
 } configuration;
 
+/****************************************/
+// DAQmx Configuration Options
+/****************************************/
 
+float64 minVal = -10.0; // The minimum value, in units, that you expect to generate.
+float64 maxVal = 10.0; // The maximum value, in units, that you expect to generate.
+float64 timeout = -1; // The amount of time, in seconds, to wait for the function to read the sample(s).
+// TODO: NOT RESPONDING
+bool32 dataLayout = DAQmx_Val_GroupByScanNumber; // Specifies how the samples are arranged, either interleaved or noninterleaved. Options: DAQmx_Val_GroupByChannel, DAQmx_Val_GroupByScanNumber
+
+/****************************************/
+// Global variables
+/****************************************/
+
+unsigned int reps = 1;   // number of repetitions of the stimulation protocol
+int currrep = 1;         // current repetition
+unsigned int isi = 0;   // inter-stimulus interval in seconds
+char *suffix = "";    // suffix for the output data file
+
+char *configFile = "geko.ini";  // name of the configuration file
+char *stimFile = "default.stim";    // name of the stimulation protocol file
+char *FileName = NULL;     // name of the output data file to be created
+char *user_prompt = NULL;   // command used to run this program
+int ai_factor = 1;
+double ao_factor = 1.0;
 configuration config;
+int silence = 0;
+const char *devName;
 
-// FUNCTION PROTOTYPES
+/****************************************/
+// Function prototypes
+/****************************************/
+
 char *getUniqueFileName(); // function to generate a unique file name
+char *getCommand(int argc, char *argv[]); // function to concatenate the command line arguments into a single string
 int handleArgs(int argc, char *argv[]); // function to handle the command line arguments
-void checkPars(); // function to check the command line args against defaults
-void printPars(); // function to print the parameters used by the program
-void printHelp(); // function to print the help message
-void readwriteFinite(); 
-void createDataStore();
-void print_image(FILE *fptr);
-static int handler(void* user, const char* section, const char* name,
-                   const char* value);
+static int handler(void* user, const char* section, const char* name, const char* value);
 int readConfigFile();
+void setConversionFactors(); // function to set the conversion factors for the DAQ
+void printPars(); // function to print the parameters used by the program
+void print_image(FILE *fptr);
+void printHelp(); // function to print the help message
+double *generateStimArray(char *stimFile); // function to generate the stimulation array
+void mkOutputDir(); // function to create the output directory
+void doSingleTask(); // function to do a single task
+void doTask();
+void mkExperimentalLog();
 
-// FUNCTION DEFINITIONS
+/****************************************/
+// Function definitions
+/****************************************/
 
-// Function to generate a unique (outut data) file name
 char *getUniqueFileName() {         //------------------------------------------
 // The first 8 characters are the date in the format YYYYMMDD.
 // The last 9 characters are the time in the format HHMMSSmmm.
@@ -93,25 +104,28 @@ char *getUniqueFileName() {         //------------------------------------------
     return fname;
 } // end of getUniqueFileName() ------------------------------------------------
 
-char *getCommand(int argc, char *argv[]) {         //------------------------------------------
-int string_length;
+char *getCommand(int argc, char *argv[]) {         //---------------------------
+// This function concatenates the command line arguments into a single string.
+    int string_length = 0;
     int i;
-for (i = 0; i < argc; i++) {
+    for (i = 0; i < argc; i++) {
         string_length = string_length + strlen(argv[i]);
     }
+    
     char *out = malloc(string_length + 1);
-
+    out[0] = '\0';
     for (i = 0; i < argc; i++) {
         strcat(out, argv[i]);
         strcat(out, " ");
     }
     return out;
-}
-// Function to handle by getopts the command line arguments
-int handleArgs(int argc, char *argv[]) {    //----------------------------------
-    int c;                                   // variable to store the option
+} // end of getCommand() -------------------------------------------------------
 
-    while ((c = getopt(argc, argv, "c:s:h:r:i:x:")) != -1) {
+int handleArgs(int argc, char *argv[]) {    //----------------------------------
+// Function to handle by getopts the command line arguments
+    int c;     // variable to store the option
+
+    while ((c = getopt(argc, argv, "c:s:hr:i:x:q")) != -1) {
         switch (c) {
             case 'c':
                 configFile = optarg;
@@ -131,6 +145,9 @@ int handleArgs(int argc, char *argv[]) {    //----------------------------------
             case 'x':
                 suffix = optarg;
                 break;
+            case 'q':
+                silence = 1;
+                break;
             case '?':
                 if (optopt == 'r' || optopt == 'i' || optopt == 's')
                     fprintf(stderr, "Option -%c requires an argument.", optopt);
@@ -148,24 +165,34 @@ int handleArgs(int argc, char *argv[]) {    //----------------------------------
     return optind;
 } // end of handleArgs() -------------------------------------------------------
 
-static int handler(void* user, const char* section, const char* name,
-                   const char* value)
-{
+static int handler(void* user, const char* section, const char* name, const char* value)  { 
+// This function is called by the ini parser for each key-value pair in the config file
     configuration* pconfig = (configuration*)user;
     char *eptr;
-    const char *devName;
+    char *aiCh;
+    char *aoCh;
+
     #define MATCH(s, n) strcmp(section, s) == 0 && strcmp(name, n) == 0
     if (MATCH("hardware", "device")) {
-        pconfig->device = value;
-        devName = value;
+        devName = strdup(value);
+        pconfig->device = devName;
+        
     } else if (MATCH("hardware", "aichans")) {
-        pconfig->aiCh = value;
+        aiCh = strdup(devName);
+        strcat(aiCh,"/");
+        strcat(aiCh,strdup(value));
+        pconfig->aiCh = aiCh;
     } else if (MATCH("hardware", "aochans")) {
-        pconfig->aoCh = value;
+        aoCh = strdup(devName);
+        strcat(aoCh,"/");
+        strcat(aoCh,strdup(value));
+        pconfig->aoCh = aoCh;
     } else if (MATCH("hardware", "srate")) {
         pconfig->srate = atoi(value);
+    } else if (MATCH("hardware", "nrchans")) {
+        pconfig->nA = atoi(value);
     } else if (MATCH("amplifier", "mode")) {
-        pconfig->mode = value;
+        pconfig->mode = strdup(value);
     } else if (MATCH("amplifier", "ai_factor_cc")) {
         pconfig->ai_factor_cc = atoi(value);
     } else if (MATCH("amplifier", "ai_factor_vc")) {
@@ -178,51 +205,34 @@ static int handler(void* user, const char* section, const char* name,
         return 0;  /* unknown section/name, error */
     }
     return 1;
-}
+} // end of handler() ----------------------------------------------------------
 
-// Read the configuration file
-int readConfigFile() {                 //-------------------------------------
-
+int readConfigFile() {                 //---------------------------------------
+// This function reads the configuration file and stores the values in the config struct
     if (ini_parse(configFile, handler, &config) < 0) {
         printf("Can't load '%s'\n", configFile);
         return 1;
     }
-
     return 0;
-} // end of readConfigFile() --------------------------------------------------
+} // end of readConfigFile() ---------------------------------------------------
 
-
-// Function to check the command line arguments against defaults
-void checkPars() {                       //-------------------------------------
-    if (config.srate <= 0) {
-        config.srate = 15000;     // 15 kHz default sampling rate
-    }
-    if (config.nA <= 1) {
-        config.nA = 1;                 // 1 analog input channel by default
-        config.aiCh = "Dev1/ai0:1";
-        config.aoCh = "Dev1/ao0";
-    } else if (config.nA >= 2) {
-        config.nA = 2;
-        config.aiCh = "Dev1/ai0:3";
-        config.aoCh = "Dev1/ao0:1";
-    }
-    if (stimFile == NULL) {
-        stimFile = "default.stim";
-    }
-    if (config.mode == "cc") {
+void setConversionFactors() {           //---------------------------------------
+// This function sets the conversion factors for the analog input and output
+    if (strcmp(config.mode, "cc") == 0) {
         ai_factor = config.ai_factor_cc;
         ao_factor = config.ao_factor_cc;
-    } else if (config.mode == "vc") {
+    } else if (strcmp(config.mode, "vc") == 0) {
         ai_factor = config.ai_factor_vc;
         ao_factor = config.ao_factor_vc;
+    } else {
+        printf("Unknown amplifier mode: %s\n", config.mode);
+        exit(1);
     }
-    
-} // end of checkPars() --------------------------------------------------------
+} // end of setConversionFactors() --------------------------------------------
 
-
-FILE *fptr = NULL;
-// Function to print the parameters used by the program, separated by ========
 void printPars() {                       //-------------------------------------
+// Function to print the parameters used by the program, separated by ========
+    FILE *fptr = NULL;
     printf("\n");
     fptr = fopen("../img/logo.txt", "r");
     print_image(fptr);
@@ -238,17 +248,14 @@ void printPars() {                       //-------------------------------------
     printf("\n");
 } // end of printPars() --------------------------------------------------------
 
-
-void print_image(FILE *fptr)
-{
+void print_image(FILE *fptr) {         //-------------------------------------
     char read_string[128];
-
     while(fgets(read_string,sizeof(read_string),fptr) != NULL)
         printf("%s",read_string);
-}
+} // end of print_image() -----------------------------------------------------
 
-// Function to print the help message
 void printHelp() {                       //-------------------------------------
+// Function to print the help message
     printf("%s\n\n", GREET);
     printf(" Usage: geko [options] \n");   
     printf(" Options: \n");    
@@ -257,74 +264,81 @@ void printHelp() {                       //-------------------------------------
     printf(" -r <Number of stimulus repetitions> \n");
     printf(" -i <Inter stimuli interval> \n");
     printf(" [-x] <Suffix for output file> \n");
+    printf(" [-q] Silence mode \n");
 } // end of printHelp() --------------------------------------------------------
 
-
-float64 stimArray[15000];
-// Function to generate the stimulation array, based on the filename supplied
-void generateStimArray() {      //-----------------------------------------------
+double *generateStimArray(char *stimFile) {      //-----------------------------------------------
     // stimFile
     // stimArray = stimgen(stim);
-
-    // Assign to the variable stimArray the value of zero vector
-    // of length 1 s * sampling rate
+    static double outArray[15000];
     printf("Generating stimulation array...\n");
     int i;
     for (i = 0; i < config.srate; i++)
     {
-        stimArray[i] = 0;
+        outArray[i] = 0;
     }    
-    int32 size_stim = sizeof(stimArray)/sizeof(stimArray[0]);
+
+    int size_stim = sizeof(outArray)/sizeof(outArray[0]);
     // Multiply stimArray by ao_factor
     for (i = 0; i < size_stim; i++)
     {
-        stimArray[i] = stimArray[i]*ao_factor;
+        outArray[i] = outArray[i]*ao_factor;
     }
-    printf("Stimulation array size: %d\n", size_stim);
-    return;
+    return outArray;
 
 } // end of generateStimArray() -------------------------------------------------
 
+void mkOutputDir() {                     //---------------------------------------
+// This function creates the output directory
+    FileName = getUniqueFileName();
+    char *extraFileName = malloc(2*strlen(FileName)+3);
+    sprintf(extraFileName,"%s/.%s", FileName,FileName);
 
-// Generalized function to interact with DAQ
-void readwriteFinite() {
+    int result = mkdir(FileName, 0777);
+    if (result != 0)
+    {
+        printf("Error creating directory\n");
+        return;
+    }
+
+    result = mkdir(extraFileName, 0777);
+    if (result != 0)
+    {
+        printf("Error creating directory\n");
+        return;
+    }
+} // end of mkOutputDir() -----------------------------------------------------
+
+void doSingleTask() {                    //---------------------------------------
+// This function performs a single task, i.e. a single stimulation protocol
+    float64 data[300000]; // Data array to be written
+    char *fullFileName = malloc(200);
+    FILE *fp;
+    int error=0;
+    char errBuff[2048]={'\0'};
     // Variables associated with the NIDAQ tasks
     TaskHandle AITaskHandle=0, AOTaskHandle=0;
     int32 read; // How many samples have been read
     int32 written; // How many samples have been written
     int32 sampsPerChan = 15000; //sizeof(stimArray)/sizeof(stimArray[0]); // Number of samples to generate/acquire per channel.
     int32 size_data = sampsPerChan*config.nA*2; // Size of the data array
-    //float64 data[size_data]; // Data array to be written
-    float64 data[300000]; // Data array to be written
-    FILE *fp, *fs, *ft;
-    int error=0;
-    char errBuff[2048]={'\0'};
-    char ch, notes[1000];
-   
-    FileName = getUniqueFileName();
-     
+    
+    // Generate stimulation array
+    double *stimArray = generateStimArray(stimFile);
+    printf("Performing repetition %d of %d...\n", currrep, reps);
     DAQmxErrChk (DAQmxCreateTask("",&AITaskHandle));
     DAQmxErrChk (DAQmxCreateAIVoltageChan(AITaskHandle,config.aiCh,"",DAQmx_Val_RSE,minVal,maxVal,DAQmx_Val_Volts,NULL));
     DAQmxErrChk (DAQmxCfgSampClkTiming(AITaskHandle,NULL,config.srate,DAQmx_Val_Rising,DAQmx_Val_FiniteSamps,sampsPerChan));
-  
     DAQmxErrChk (DAQmxCreateTask("",&AOTaskHandle));   
     DAQmxErrChk (DAQmxCreateAOVoltageChan(AOTaskHandle,config.aoCh,"",minVal,maxVal,DAQmx_Val_Volts,NULL));
     DAQmxErrChk (DAQmxCfgSampClkTiming(AOTaskHandle,"",config.srate,DAQmx_Val_Rising,DAQmx_Val_FiniteSamps,sampsPerChan));
-
     // Connect AO start to AI start
     DAQmxErrChk (DAQmxCfgDigEdgeStartTrig(AOTaskHandle, "ai/StartTrigger", DAQmx_Val_Rising));
-
     // Arm the AO task
     DAQmxErrChk (DAQmxWriteAnalogF64(AOTaskHandle,sampsPerChan,FALSE,timeout,dataLayout, stimArray, &written, NULL));
-
-    // Start the AI task
-    printf("Performing the task...\n");
     DAQmxErrChk (DAQmxStartTask(AOTaskHandle));
     DAQmxErrChk (DAQmxStartTask(AITaskHandle));
-    
     DAQmxErrChk (DAQmxReadAnalogF64(AITaskHandle,sampsPerChan,timeout,dataLayout,data,size_data,&read,NULL));
-    // int32 DAQmxReadAnalogF64 (TaskHandle taskHandle, int32 numSampsPerChan, float64 timeout, bool32 fillMode, float64 readArray[], uInt32 arraySizeInSamps, int32 *sampsPerChanRead, bool32 *reserved);
-
     DAQmxErrChk (DAQmxClearTask(AOTaskHandle));
     DAQmxErrChk (DAQmxClearTask(AITaskHandle));
 
@@ -334,68 +348,13 @@ void readwriteFinite() {
     {
         data[i] = data[i]*ai_factor;
     }
-
-    char *fullFileName = malloc(strlen(FileName)+strlen(FILESUFFIX)+1);
-    sprintf(fullFileName,"%s%s",FileName,FILESUFFIX);
-     printf("Made it here!\n");
+    
+    sprintf(fullFileName,"%s/%s_%i%s",FileName,FileName,currrep,FILESUFFIX);
     fp = fopen(fullFileName, "wb");
-    fwrite(&data, sizeof(data), 1, fp);
+    fwrite(data, sizeof(data), 1, fp);
     fclose(fp);
-
-    sprintf(fullFileName,".%s",FileName);
-    int result = mkdir(fullFileName, 0777);
-    if (result == 0)
-    {
-        fs = fopen(configFile, "r");
    
-        sprintf(fullFileName,".%s/%s",FileName,configFile);
-        ft = fopen(fullFileName, "w");
-        ch = fgetc(fs);
-        while (ch != EOF)
-        {
-            fputc(ch, ft);
-            ch = fgetc(fs);
-        }
-        fclose(fs);
-        fclose(ft);
-
-        fs = fopen(stimFile, "r");
-    
-        sprintf(fullFileName,".%s/%s",FileName,stimFile);
-        ft = fopen(fullFileName, "w");
-        ch = fgetc(fs);
-        while (ch != EOF)
-        {
-            fputc(ch, ft);
-            ch = fgetc(fs);
-        }
-        fclose(fs);
-        fclose(ft);
-
-    
-
-        printf("Do you have any experimental notes to add? (Press SPACE ENTER to continue)\n");
-        // fgets(notes);
-        // Get user input
-        scanf("%s", notes);
-      
-        sprintf(fullFileName,".%s/notes.txt",FileName);
-        ft = fopen(fullFileName, "w+");
-        fputs("[string used]\n", ft);
-
-        fputs(user_prompt, ft);
-        fputs("\n\n", ft);
-        fputs("[experimental notes]\n", ft);
-        fputs(notes, ft);
-    }
-    else
-    {
-        printf("Error creating directory\n");
-        return;
-    }
-
-
-    printf("Task completed.\n");
+    printf("Data saved to %s\n\n", fullFileName);
 
     Error:
 	if( DAQmxFailed(error) )
@@ -417,17 +376,69 @@ void readwriteFinite() {
 		AOTaskHandle = 0;
 	}
 	if( DAQmxFailed(error) )
-		printf("DAQmx Error: %s\n",errBuff);
+		printf("DAQmx Error: %s\n",errBuff);    
+      
+} // end of doSingleTask() ----------------------------------------------------
 
-    return;
-}
+void mkExperimentalLog() {               //---------------------------------------
+    FILE  *fs1, *ft1, *fs2, *ft2, *ft = NULL;
+    char ch1, ch2, notes[1000];
+    char *fullFileName = malloc(300);
 
-void createDataStore() {
-  
-    return;
+    fs1 = fopen(configFile, "r");
+    sprintf(fullFileName,"%s/.%s/%s",FileName,FileName,configFile);
+    
+    ft1 = fopen(fullFileName, "w");
+    ch1 = fgetc(fs1);
+    
+    while (ch1 != EOF)
+    {
+        fputc(ch1, ft1);
+        ch1 = fgetc(fs1);
+    }
+    fclose(fs1);
+    fclose(ft1);
+
+    fs2 = fopen(stimFile, "r");
+    sprintf(fullFileName,"%s/.%s/%s",FileName,FileName,stimFile);
+    ft2 = fopen(fullFileName, "w");
+    ch2 = fgetc(fs2);
+    while (ch2 != EOF) 
+    {
+        fputc(ch2, ft2);
+        ch2 = fgetc(fs2);
+    }
+    fclose(fs2);
+    fclose(ft2);
+
+    printf("Do you have any experimental notes to add? (Press ENTER to continue)\n");
+    // Get user input
+    fgets(notes, 1000, stdin);
+
+    sprintf(fullFileName,"%s/.%s/notes.txt",FileName,FileName);
+    ft = fopen(fullFileName, "w+");
+    fputs("[string used]\n", ft);
+
+    fputs(user_prompt, ft);
+    fputs("\n\n", ft);
+    fputs("[experimental notes]\n", ft);
+    fputs(notes, ft);
 
 
-}
+    printf("Task completed.\n");
 
+} // end of mkExperimentalLog() -----------------------------------------------
 
-
+void doTask() {                          //---------------------------------------
+// This function performs the task, i.e. the stimulation protocol
+    setConversionFactors();
+    mkOutputDir();
+    int i;
+    for (i = 0; i < reps; i++)
+    {
+        currrep = i + 1;
+        doSingleTask();
+        sleep(isi);
+    }
+    mkExperimentalLog();
+} // end of doTask() ----------------------------------------------------------
