--- src/.20230221_182624/default.stim
@@ -1,2 +1,0 @@
-1   1   0   0   0   0   0   0   0   0   0   1
-1   1   0   0   0   0   0   0   0   0   0   2
