--- src/.20230221_182624/notes.txt
@@ -1,4 +1,0 @@
-[string used]
-õ.ñ‚SU./geko -r 2 -i 1 
-[experimental notes]
-f
