+++ src/20230222_120729/.20230222_120729/default.stim
@@ -1,0 +1,2 @@
+1   1   0   0   0   0   0   0   0   0   0   1
+1   1   0   0   0   0   0   0   0   0   0   2
