+++ src/20230222_120729/.20230222_120729/geko.ini
@@ -1,0 +1,51 @@
+; Sample config file for geKo
+
+; This is a comment
+
+
+
+; Data acquisition hardware configuration
+[hardware]
+device = Dev1       ; Device name, e.g. Dev1 (as in Dev1/ai0)
+
+ai0_cal_offset = 0      ; Offset for calibration, default 0
+ai0_cal_scale  = 1      ; Factor for calibration, default 1
+ai1_cal_offset = 0      ; Offset for calibration, default 0
+ai1_cal_scale  = 1      ; Factor for calibration, default 1
+
+ao0_cal_offset = 0      ; Offset for calibration, default 0
+ao0_cal_scale  = 1      ; Factor for calibration, default 1
+ao1_cal_offset = 0      ; Offset for calibration, default 0
+ao1_cal_scale  = 1      ; Factor for calibration, default 1
+
+nrchans = 1          ; Number of channels, default 1
+aichans = ai0:1       ; Analog input channels, default ai0 ai1
+aochans = ao0           ; Analog output channels, default ao0
+
+srate =  15000          ; Sample rate [Hz], default 15000
+
+;-------------------------------------------------------------------------------
+; Amplifier configuration (e.g. for MultiClamp 700B)
+[amplifier]
+
+mode = cc               ; Amplifier mode, cc or vc
+ai_factor_cc = 20          ; Amplifier gain in current-clamp
+ai_factor_vc = 400         ; Amplifier gain in voltage-clamp
+
+ao_factor_cc = 0.00025     ; pA/V of the Amplifier's command in current-clamp
+ao_factor_vc = 0.05        ; mV/V of the Amplifier's command in current-clamp
+
+
+
+;-------------------------------------------------------------------------------
+
+
+
+[protocol]             ; Protocol configuration
+version=26              ; IPv6
+
+[user]
+name = Aldio Bob Smith       ; Spaces around '=' are stripped
+email = bob@smith.com  ; And comments (like this) ignored
+active = true          ; Test a boolean
+pi = 3.14159           ; Test a floating point number
