+++ src/20230222_120729/.20230222_120729/notes.txt
@@ -1,0 +1,5 @@
+[string used]
+./geko -r 2 -i 1 
+
+[experimental notes]
+Yes!
