+++ src/geko.ini
@@ -6,7 +6,7 @@
 
 ; Data acquisition hardware configuration
 [hardware]
-device = Dev1/       ; Device name, e.g. Dev1 (as in Dev1/ai0)
+device = Dev1       ; Device name, e.g. Dev1 (as in Dev1/ai0)
 
 ai0_cal_offset = 0      ; Offset for calibration, default 0
 ai0_cal_scale  = 1      ; Factor for calibration, default 1
@@ -18,7 +18,7 @@ ao0_cal_scale  = 1      ; Factor for calibration, default 1
 ao1_cal_offset = 0      ; Offset for calibration, default 0
 ao1_cal_scale  = 1      ; Factor for calibration, default 1
 
-nrChannels = 1          ; Number of channels, default 1
+nrchans = 1          ; Number of channels, default 1
 aichans = ai0:1       ; Analog input channels, default ai0 ai1
 aochans = ao0           ; Analog output channels, default ao0
